Require services to opt-in
Services that want to be exposed via operatord and operatorctl now have
to opt-in explicity using the "operator.enabled" service level option.
This allows running protoc-gen-operator{d,ctl} on a directory of proto
files where not all services should necessarily be exposed as chat
commands.

fbshipit-source-id: bc6be3b
