+++ handler.go
@@ -7,8 +7,6 @@ import (
 	"strings"
 	"unicode"
 
-	"github.com/sr/operator/generator"
-
 	"golang.org/x/net/context"
 )
 
@@ -69,8 +67,8 @@ func (h *handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		&Request{
 			Call: &Call{
 				// TODO(sr) multi package support
-				Service: fmt.Sprintf("%s.%s", h.pkg, generator.Camelize(matches[1], "-")),
-				Method:  generator.Camelize(matches[2], "-"),
+				Service: fmt.Sprintf("%s.%s", h.pkg, Camelize(matches[1], "-")),
+				Method:  Camelize(matches[2], "-"),
 				Args:    args,
 			},
 			Otp:      otp,
