+++ testing/integration_test.sh
@@ -7,19 +7,30 @@ cd "$(dirname "$0")/.."
 # Install protoc-gen-operatord and protoc-gen-operatorctl
 go install -v github.com/sr/operator/cmd/...
 
-# Generater operatorctl code from testing.proto using protoc-gen-operatorctl
+# Generater operatorctl (client) and operatord (server) code from protobuf files
+mkdir "${TEST_TMPDIR}/operatorctl" "${TEST_TMPDIR}/operatord"
 protoc \
     -I"$(brew --prefix protobuf)/include" \
     -Itesting \
     -I. \
-    --operatorctl_out="import_path=github.com/sr/operator/testing:${TEST_TMPDIR}" \
+    --operatorctl_out="import_path=github.com/sr/operator/testing:${TEST_TMPDIR}/operatorctl" \
+    --operatord_out="import_path=github.com/sr/operator/testing:${TEST_TMPDIR}/operatord" \
     testing/*.proto
 
-# Compile generated code into operatorctl and place it on the PATH
+# Compile generated client code into operatorctl and place it on the PATH
 mkdir "${TEST_TMPDIR}/bin"
-go build -o "${TEST_TMPDIR}/bin/operatorctl" "${TEST_TMPDIR}/main-gen.go"
+go build -o "${TEST_TMPDIR}/bin/operatorctl" "${TEST_TMPDIR}/operatorctl/main-gen.go"
 PATH="${TEST_TMPDIR}/bin:$PATH"
 export PATH
+
+# Compile generated server code and discard it; this test only cares about
+# whether the code compiles at this time
+cat <<EOS > "${TEST_TMPDIR}/operatord/main.go"
+package main
+func main() {}
+EOS
+go build -o "${TEST_TMPDIR}/bin/operatord" "${TEST_TMPDIR}/operatord/main.go"
+rm -f "${TEST_TMPDIR}/bin/operatord"
 
 TEST_log="${TEST_TMPDIR}/log"
 touch "$TEST_log"
@@ -33,6 +44,9 @@ fail() {
 operatorctl -help &>"$TEST_log"
 if ! grep -qs "Usage: operatorctl" "$TEST_log"; then
 	fail "program help message did not include usage line"
+fi
+if grep -qa "private-service" "$TEST_log"; then
+    fail "program help includes help for private service"
 fi
 
 if operatorctl pinger boomtown &>"$TEST_log"; then
