+++ testing/testing.proto
@@ -7,8 +7,13 @@ package testing;
 option go_package = "operatortesting";
 
 service Pinger {
+	option (operator.enabled) = true;
 	rpc Ping(PingRequest) returns (operator.Response);
 	rpc PingPong(PingRequest) returns (operator.Response);
+}
+
+service PrivateService {
+	rpc Ping(PingRequest) returns (operator.Response);
 }
 
 message PingRequest {
